pyproject.toml
[project]
name = "medicalquestionnaire"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "unicorn>=2.1.4",
]
